refactor: extract _count_nodes, _find_node, _build_depth_map to multiverse/utils
