+++ server/__init__.py
@@ -17,6 +17,7 @@ import causality
 from agents.agent import Agent
 from multiverse.generator import generate_node_hierarchy
 from multiverse.node import SpatialNode
+from multiverse.utils import _build_depth_map, _count_nodes, _find_node
 from puzzles.engine import PuzzleEngine
 import persistence
 
@@ -145,30 +146,6 @@ def _node_to_dict(node: SpatialNode) -> dict:
         "properties": node.properties,
         "children": [_node_to_dict(c) for c in node.children],
     }
-
-
-def _count_nodes(node: SpatialNode) -> int:
-    return 1 + sum(_count_nodes(c) for c in node.children)
-
-
-def _find_node(root: SpatialNode, name: str) -> SpatialNode | None:
-    if root.name == name:
-        return root
-    for child in root.children:
-        found = _find_node(child, name)
-        if found:
-            return found
-    return None
-
-
-def _build_depth_map(node: SpatialNode, depth: int = 0,
-                     result: dict | None = None) -> dict[str, int]:
-    if result is None:
-        result = {}
-    result[node.id] = depth
-    for child in node.children:
-        _build_depth_map(child, depth + 1, result)
-    return result
 
 
 def _rebuild(params: dict) -> tuple[SpatialNode, int, int, int, int]:
